Create Boid update Function to Move Boids
This commit adds the function `update` in the `boid` namespace to
update a boid in a single time step.

The docstring of `update` explains the purpose of that function. Or
more interest here is the intent of that function: Having an update
function that returns an updated boid object, combined with making
flocks iterable, allows patterns like this to update the flock for a
single time step:

```clojure
(let [boids (boids flock)
      create-boid (...)
      f (fn [boid]
          (update create-boid boid flock))]
  (map f boids)
```