Create Update Loop
I also decomplected `boid/update` and `create-boid`, because they have
do not need to be conjoined.

In fact, a simpler representation of boids may be that they are simply
maps of `#{::position ::velocity}`, and maybe `::update-f`. I don't
need a record to contain that information or to bind an update
function to that data. All I need is an update function that take a
single arg, the boid data structure itself.