Rewrite Boids Using clojure.spec
I realized that my previous boids model -- using a protocol --
unnecessarily bundled boid state and behavior. Most boids can simply
be represented as a map of position and velocity, and *maybe* a
behavior function.

Clojure does great at working with maps, so why was I creating a
custom Boid type and losing out on all of Clojure's awesomeness?

This approach instead defines a boid using clojure.spec. A boid is
simply a position and a velocity. (I can add a behavior function in
later if I need it.)

I also defined two simple operations on boids: First, adding a vector
to a boid is assumed to mean adding an acceleration to the boid's
velocity. I made this assumption because adding the vector to the
boid's position would allow boids to jump in space.

Second, a `move` function was created that is basically a slimmed-down
version of the `update` function previously. The previous function
unnecessarily bundled running the update function with the actual
state-changing operation on the boid. `move` only does the
state-changing; this leaves `move` simpler, in that it involves one
and only one thing now.