Add Concrete Implementation of Vector
This change adds a concrete implementation of Vector.

More interesting is that this change fully provides spec stuff for
Vector. The proved very enlightening because it highlighted a weakness
in the design of Vector: doubles.

1. Doubles have a few weird values: positive infinity, negative
   infinity, and NaN. My `defspec` tests demonstrated that my Vector
   implementation was not correctly handling these values. I ended up
   just excluding those values from the allowed values in Vectors.
2. My Vector invariants did not correctly handle double rounding
   error. See, e.g., the `magnitude` test.

Also, my `defspec` tests also found that I initially used
`Math/arctan2` incorrectly, reversing the order of the paramters. Spec
rocks.